refactor prim tests into new file, remove param and prims without proper types
