+++ frontend/test/resolution/CMakeLists.txt
@@ -64,6 +64,7 @@ comp_unit_test(testReflection)
 comp_unit_test(testResolve)
 comp_unit_test(testResolverVerboseErrors)
 comp_unit_test(testReturnTypes)
+comp_unit_test(testRuntimePrimitives)
 comp_unit_test(testScopeResolve)
 comp_unit_test(testScopeTypes)
 comp_unit_test(testSplitInit)
